Move meta db to separate crate
