+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = [ "ingest", "store", "mobile_rewards", "status", "mobile_verifier", "entropy",
+members = [ "ingest", "db_store", "file_store", "mobile_rewards", "status", "mobile_verifier", "entropy",
 "lora_verifier", "poc_iot_injector"]
