+++ entropy/src/entropy_generator.rs
@@ -1,5 +1,6 @@
 use crate::{Error, Result};
 use chrono::Utc;
+use file_store::file_sink;
 use futures::TryFutureExt;
 use helium_proto::EntropyReportV1;
 use jsonrpsee::{
@@ -7,7 +8,6 @@ use jsonrpsee::{
     http_client::{HttpClient, HttpClientBuilder},
     rpc_params,
 };
-use poc_store::file_sink;
 use serde::{Deserialize, Serialize};
 use serde_json::json;
 use std::env;
