+++ entropy/src/main.rs
@@ -1,8 +1,8 @@
 use chrono::Duration;
 use clap::Parser;
+use file_store::{file_sink, file_upload, FileType};
 use futures_util::TryFutureExt;
 use poc_entropy::{entropy_generator::EntropyGenerator, server::ApiServer, Error, Result};
-use poc_store::{file_sink, file_upload, FileType};
 use std::path;
 use tokio::{self, signal};
 use tracing_subscriber::{layer::SubscriberExt, util::SubscriberInitExt};
