+++ file_store/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "poc-store"
+name = "file-store"
 version = "0.1.0"
 edition = "2021"
-description = "Ingest Store Library"
-authors = ["Marc Nijdam <marc@nova-labe.com>"]
+description = "Ingest File Store Library"
+authors = ["Marc Nijdam <marc@nova-labs.com>"]
 license = "Apache-2.0"
 
 [dependencies]
@@ -11,7 +11,6 @@ dotenv = "0"
 clap = {version = "3", features = ["derive"]}
 serde =  {version = "1", features=["derive"]}
 serde_json = "1"
-sqlx = {version = "0", features = ["postgres", "uuid", "decimal", "chrono", "migrate", "macros", "runtime-tokio-rustls"]}
 thiserror = "1"
 tokio = { version = "1", default-features=false, features=["fs", "rt-multi-thread"] }
 tokio-util = "0"
