+++ file_store/src/cli/bucket.rs
