+++ file_store/src/cli/dump.rs
