+++ file_store/src/cli/info.rs
