+++ file_store/src/cli/mod.rs
