+++ file_store/src/lib.rs
@@ -9,7 +9,6 @@ pub mod heartbeat;
 pub mod lora_beacon_report;
 pub mod lora_valid_poc;
 pub mod lora_witness_report;
-pub mod meta;
 pub mod speedtest;
 pub mod traits;
 
@@ -17,7 +16,6 @@ pub use error::{Error, Result};
 pub use file_info::{FileInfo, FileType};
 pub use file_sink::{FileSink, FileSinkBuilder};
 pub use file_store::FileStore;
-pub use meta::{MetaError, MetaValue};
 
 use bytes::BytesMut;
 use chrono::{DateTime, NaiveDateTime, Utc};
