+++ ingest/Cargo.toml
@@ -35,7 +35,7 @@ helium-proto = { git = "https://github.com/helium/proto", branch = "master", fea
   "services",
 ] }
 helium-crypto = {git = "https://github.com/helium/helium-crypto-rs", tag="v0.4.4", features=["sqlx-postgres"]}
-poc-store = { path = "../store" }
+file-store = { path = "../file_store" }
 poc-metrics = { path = "../metrics" }
 metrics = "0.20"
 metrics-exporter-prometheus = "0.11.0"
