+++ lora_verifier/Cargo.toml
@@ -30,7 +30,7 @@ axum = "0"
 async-compression = {version = "0", features = ["tokio", "gzip"]}
 helium-proto = {git = "https://github.com/helium/proto", branch = "master", features = ["services"]}
 helium-crypto = {git = "https://github.com/helium/helium-crypto-rs", tag="v0.4.4"}
-poc-store = { path = "../store" }
+file-store = { path = "../file_store" }
 async-trait = "*"
 lazy_static = "*"
 once_cell = "1"
