+++ lora_verifier/src/error.rs
@@ -27,7 +27,7 @@ pub enum Error {
     #[error("crypto error")]
     Crypto(#[from] helium_crypto::Error),
     #[error("store error")]
-    Store(#[from] poc_store::Error),
+    Store(#[from] file_store::Error),
     #[error("not found")]
     NotFound(String),
     #[error("base64 decode error")]
