+++ lora_verifier/src/lib.rs
@@ -10,7 +10,7 @@ pub mod runner;
 pub mod traits;
 
 pub use error::{Error, Result};
-pub use poc_store::datetime_from_epoch;
+pub use file_store::datetime_from_epoch;
 use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
 use std::io;
