+++ mobile_rewards/src/cli/generate.rs
@@ -2,9 +2,9 @@ use crate::{
     cli::print_json, follower::FollowerService, subnetwork_rewards::SubnetworkRewards, Result,
 };
 use chrono::NaiveDateTime;
+use file_store::{datetime_from_naive, FileStore};
 use helium_crypto::PublicKey;
 use helium_proto::SubnetworkReward as ProtoSubnetworkReward;
-use poc_store::{datetime_from_naive, FileStore};
 use serde_json::json;
 
 /// Generate poc rewards
