+++ mobile_rewards/src/server.rs
@@ -12,6 +12,7 @@ use crate::{
     Error, Result,
 };
 use chrono::{Duration, Utc};
+use file_store::FileStore;
 use futures::try_join;
 use helium_proto::{
     blockchain_txn::Txn,
@@ -22,7 +23,6 @@ use helium_proto::{
     BlockchainTokenTypeV1, BlockchainTxn, BlockchainTxnSubnetworkRewardsV1, Message,
 };
 use poc_metrics::record_duration;
-use poc_store::FileStore;
 use sqlx::{Pool, Postgres};
 use std::env;
 use tokio::time;
