+++ mobile_verifier/src/error.rs
@@ -23,13 +23,13 @@ pub enum Error {
     #[error("crypto error")]
     Crypto(#[from] helium_crypto::Error),
     #[error("store error")]
-    Store(#[from] poc_store::Error),
+    Store(#[from] file_store::Error),
     #[error("not found")]
     NotFound(String),
     #[error("transaction error")]
     TransactionError(String),
     #[error("meta error")]
-    MetaError(#[from] poc_store::MetaError),
+    MetaError(#[from] db_store::MetaError),
     #[error("join error")]
     JoinError(tokio::task::JoinError),
 }
