+++ mobile_verifier/src/server.rs
@@ -4,9 +4,10 @@ use crate::{
     subnetwork_rewards::SubnetworkRewards,
 };
 use chrono::{DateTime, Duration, NaiveDateTime, Utc};
+use db_store::MetaValue;
+use file_store::{file_sink, file_upload, FileStore, FileType};
 use futures_util::TryFutureExt;
 use helium_proto::services::{follower, Endpoint, Uri};
-use poc_store::{file_sink, file_upload, FileStore, FileType, MetaValue};
 use sqlx::postgres::PgPoolOptions;
 use tokio::{select, time::sleep};
