+++ status/src/gateway.rs
@@ -1,7 +1,7 @@
 use crate::{Error, Result};
 use chrono::{DateTime, Utc};
+use file_store::datetime_from_epoch;
 use helium_crypto::PublicKey;
-use poc_store::datetime_from_epoch;
 use serde::{Deserialize, Serialize};
 use std::cmp::min;
