+++ status/src/lib.rs
@@ -5,7 +5,7 @@ pub mod meta;
 pub mod server;
 
 pub use error::{Error, Result};
-pub use poc_store::datetime_from_epoch;
+pub use file_store::datetime_from_epoch;
 
 use sqlx::{postgres::PgPoolOptions, Pool, Postgres};
